Fixes module include / xcode
